requirements.txt
Flask
gspread
google-api-python-client
google-auth-oauthlib
python-dotenv
gunicorn
psycopg2-binary
SQLAlchemy